Move to more vite frontend correctness
Problem
-----

the goal was to get rid of the pico gem

Solution
-----

more vite config so that we're doing assets and images and css correctly with vite.

Changes
-----

* move frontend stuff into app/frontend
* update links and such
* add stylelint
* clean up css to be dart sass compliant
