+++ Gemfile
@@ -5,8 +5,8 @@ ruby '3.1.3'
 
 gem 'acts-as-taggable-on'
 gem 'addressable'
-gem 'aws-sdk-cloudfront', '~> 1.73'
-gem 'aws-sdk-s3', '~> 1'
+gem 'aws-sdk-cloudfront'
+gem 'aws-sdk-s3'
 gem 'clearance'
 gem 'dotenv-rails'
 gem 'fast_jsonapi'
