--- app/assets/javascripts/admin.js_
@@ -1,19 +1,0 @@
-// admin javascript manifest
-//= require_tree ./support
-//= require pickadate.min
-//= require jquery.timepicker
-//= require selectize
-
-$(function() {
-  $('.datepicker').pickadate()
-  $('.timepicker').timepicker()
-  $('.text_copy').bind('click', function() {
-    $(this).select()
-  });
-
-  $('#notice').bind('click', function() {
-    $(this).fadeOut()
-  })
-
-  setTimeout(function() { $('#notice').fadeOut() }, 3500)
-});
