--- app/assets/stylesheets/admin/admin.css.scss
@@ -1,22 +1,0 @@
-// /* admin */
-@use '../imports/mixins';
-@use '../imports/colors';
-@use '../imports/dimensions';
-@use '../application/loveli-icons.css.scss';
-
-header.container-fluid {
-  border-bottom: 1px solid $light-gray;
-}
-
-footer {
-  border-top: 1px solid $light-gray;
-  display: flex;
-  gap: 20px;
-  justify-content: flex-end;
-}
-
-.flash {
-  &.error, &.alert {
-    color: $dark-red;
-  }
-}
