--- app/assets/stylesheets/imports/mixins.scss
@@ -1,60 +1,0 @@
-/* mixins */
-
-@mixin opacity($opacity) {
-  filter: unquote("progid:DXImageTransform.Microsoft.Alpha(Opacity=#{round($opacity * 100)})");
-  opacity: $opacity;
-};
-
-$default-border-radius: 5px;
-
-@mixin border-radius($radius: $default-border-radius) {
-  -moz-border-radius: $radius;
-  -webkit-border-radius: $radius;
-  -khtml-border-radius: $radius;
-  -ms-border-radius: $radius;
-  border-radius: $radius;
-};
-
-@mixin border-radius-top($radius: $default-border-radius) {
-  @include border-radius-top-left($radius);
-  @include border-radius-top-right($radius);
-};
-
-@mixin border-radius-top-left($radius: $default-border-radius) {
-  -moz-border-radius-topleft: $radius;
-  -webkit-border-top-left-radius: $radius;
-  -khtml-border-top-left-radius: $radius;
-  -ms-border-top-left-radius: $radius;
-  border-top-left-radius: $radius;
-};
-
-@mixin border-radius-top-right($radius: $default-border-radius) {
-  -moz-border-radius-topright: $radius;
-  -webkit-border-top-right-radius: $radius;
-  -khtml-border-top-right-radius: $radius;
-  -ms-border-top-right-radius: $radius;
-  border-top-right-radius: $radius;
-};
-
-@mixin gradient($from, $to) {
-  /* fallback/image non-cover color */
-  background-color: $from;
-
-  /* Firefox 3.6+ */
-  background-image: -moz-linear-gradient($from, $to);
-
-  /* Safari 4+, Chrome 1+ */
-  background-image: -webkit-gradient(linear, 0% 0%, 0% 100%, from($from), to($to));
-
-  /* Safari 5.1+, Chrome 10+ */
-  background-image: -webkit-linear-gradient($from, $to);
-
-  /* Opera 11.10+ */
-  background-image: -o-linear-gradient($from, $to);
-}
-
-@mixin drop-shadow ($x: 1px, $y: 1px, $blur: 2px, $spread: 0, $color: rgba(0,0,0,0.25)) {
-  -webkit-box-shadow: $x $y $blur $spread $color;
-  -moz-box-shadow:    $x $y $blur $spread $color;
-  box-shadow:         $x $y $blur $spread $color;
-}
