+++ app/frontend/entrypoints/application.js
@@ -23,15 +23,15 @@ console.log('Vite ⚡️ Rails')
 // // Import all channels.
 // const channels = import.meta.globEager('./**/*_channel.js')
 
-import '../../assets/javascripts/onload.js';
-import '../../assets/javascripts/podcastButton.js';
+import '../javascripts/onload.js';
+import '../javascripts/podcastButton.js';
 
 // stylesheets
-import '../../assets/stylesheets/application/about.css.scss';
-import '../../assets/stylesheets/application/banner.css.scss';
-import '../../assets/stylesheets/application/loveli-icons.css.scss';
-import '../../assets/stylesheets/application/main.css.scss';
-import '../../assets/stylesheets/application/search.css.scss';
-import '../../assets/stylesheets/application/site-nav.css.scss';
-import '../../assets/stylesheets/application/tooltips.css.scss';
-import '../../assets/stylesheets/application/tracks.css.scss';
+import '../stylesheets/application/about.css.scss';
+import '../stylesheets/application/banner.css.scss';
+import '../stylesheets/application/loveli-icons.css.scss';
+import '../stylesheets/application/main.css.scss';
+import '../stylesheets/application/search.css.scss';
+import '../stylesheets/application/site-nav.css.scss';
+import '../stylesheets/application/tooltips.css.scss';
+import '../stylesheets/application/tracks.css.scss';
