+++ app/frontend/stylesheets/admin/admin_forms.css.scss
@@ -8,10 +8,12 @@
     color: colors.$medium_gray;
     font-size: small;
   }
+
   .input {
     margin-bottom: 20px;
   }
 }
+
 #error_explanation {
   color: colors.$dark-red;
 }
