+++ app/frontend/stylesheets/application/about.css.scss
