+++ app/frontend/stylesheets/application/site-nav.css.scss
@@ -1,10 +1,11 @@
-@import "../imports/colors";
+@use "../imports/colors";
 
 .site-nav__item + .site-nav__item {
   margin-left: 4px;
   padding-left: 4px;
-  border-left: 1px solid $light_gray;
+  border-left: 1px solid colors.$light_gray;
 }
+
 .site-nav {
   a {
     text-decoration: none;
