+++ app/frontend/stylesheets/application/tooltips.css.scss
@@ -7,17 +7,22 @@ a.tooltip {
   display: inline;
   position: relative;
 }
-a.tooltip:after {
+
+a.tooltip::after {
   bottom: 36px;
   left: -40px;
 }
+
 a.tooltip:hover {
   text-decoration:none;
 }
-a.tooltip:hover:after {
+
+a.tooltip:hover::after {
   background: colors.$medium_dark_gray;
+
   @include mixins.border-radius(4px);
   @include mixins.drop-shadow(0, 0, 2px, 1px, colors.$medium_dark_gray);
+
   color: #fff;
   content: attr(title);
   padding: 5px 15px;
@@ -25,6 +30,7 @@ a.tooltip:hover:after {
   z-index: 10;
   display: block;
 }
-a.tooltip.tooltip--under:after {
+
+a.tooltip.tooltip--under::after {
   bottom: -80px;
 }
