+++ app/frontend/stylesheets/imports/colors.scss
@@ -1,15 +1,13 @@
 /* colors */
 
 $black: #000;
-$dark_gray: rgba(30,0,30, 0.9);
-$medium_dark_gray: rgba(60,60,60,0.8);
+$dark_gray: rgb(30 0 30 / 90%);
+$medium_dark_gray: rgb(60 60 60 / 80%);
 $medium_gray: #666;
 $light_gray: #999;
 $xlight_gray: #aaa;
 $white: #fff;
-
 $brown_f3: #f3d4b3;
 $brown_c2: #c26d4f;
 $brown_68: #68361b;
-
-$dark_red:  #670202;
+$dark_red: #670202;
