+++ app/frontend/stylesheets/imports/dimensions.scss
@@ -3,10 +3,8 @@ $width: 100%;
 $max-width: 80%;
 $header-offset: 100px;
 $min-width: 500px;
-
 $block-margin: 20px;
 $block-padding: 20px;
-
 $screen-xs-max: 35.5em;
 $screen-sm-min: $screen-xs-max;
 $screen-sm-max: 48em;
