+++ app/frontend/stylesheets/imports/mixins.scss
@@ -1,0 +1,28 @@
+/* mixins */
+
+@mixin opacity($opacity) {
+  opacity: $opacity;
+};
+
+$default-border-radius: 5px;
+
+@mixin border-radius($radius: $default-border-radius) {
+  border-radius: $radius;
+};
+
+@mixin border-radius-top($radius: $default-border-radius) {
+  @include border-radius-top-left($radius);
+  @include border-radius-top-right($radius);
+};
+
+@mixin border-radius-top-left($radius: $default-border-radius) {
+  border-top-left-radius: $radius;
+};
+
+@mixin border-radius-top-right($radius: $default-border-radius) {
+  border-top-right-radius: $radius;
+};
+
+@mixin drop-shadow($x: 1px, $y: 1px, $blur: 2px, $spread: 0, $color: rgba(0,0,0,0.25)) {
+  box-shadow:         $x $y $blur $spread $color;
+}
