--- app/views/admin/common/_errors.slim
@@ -1,11 +1,0 @@
-.  - if @track.errors.any?
-
-    #error_explanation
-      h4
-        = pluralize(@track.errors.count, "error")
-        ' '
-        | prohibited this track from being saved:
-
-      ul
-      - @track.errors.full_messages.each do |msg|
-        li = msg
