+++ app/views/tracks/show.html.slim
@@ -1,2 +1,5 @@
-ul 
+section.banner-image
+  .banner-image--title
+    img src=vite_asset_path('images/practice_rope_words.png')
+ul
   = render '/welcome/track', track: @track
