+++ app/views/welcome/index.slim
@@ -1,8 +1,8 @@
 section.banner-image
   .banner-image--image
-    img src=image_path('practice_rope_rope.png')
+    img src=vite_asset_path('images/practice_rope_rope.png')
   .banner-image--title
-    img src=image_path('practice_rope_words.png')
+    img src=vite_asset_path('images/practice_rope_words.png')
 - unless @tracks.empty?
   ul.tracks
     - @tracks.each do |track|
