+++ config/environments/development.rb
@@ -59,7 +59,8 @@ Rails.application.configure do
   config.active_record.verbose_query_logs = true
 
   # Suppress logger output for asset requests.
-  config.assets.quiet = true
+  # we're on vite
+  # config.assets.quiet =
 
   # Raises error for missing translations.
   # config.i18n.raise_on_missing_translations = true
