+++ package.json
@@ -5,8 +5,14 @@
     "@rails/activestorage": "^8.0.200",
     "jquery": "^3.7.1",
     "sass-embedded": "^1.86.3",
+    "stylelint": "^16.19.1",
+    "stylelint-config-standard-scss": "^14.0.0",
     "vite": "^5.4.17",
     "vite-plugin-ruby": "^5.1.1"
   },
-  "dependencies": {}
+  "dependencies": {},
+  "scripts": {
+     "stylelint": "stylelint 'app/frontend/**/*.{scss,css}'",
+     "stylelint-fix": "stylelint --fix 'app/frontend/**/*.{scss,css}'"
+  }
 }
